Improve Helm Charts (#11)
* add image.tag to Charts which didn't have it yet

This allows the user of the Chart to specify which version to use. In
combination with ImagePullPolicy very handy for local development.

* cleanup helper tpls

The _helpers.tpl files were generated with an older version of the Helm CLI.
The files contained extra dashed at the end of the tags, which were completely
unnecessary.

* update README's and values.schema's with image.tag

* delete commented out code, yay Clean Code

* add documentation to multichain charts parameters

This is a first attempt to better document the Helm Chart workings and
parameters. The same layout can be used for the other charts.