build: check in ES Charts
EventStore is one of our dependencies for the registry-backend. We want to make changes to the upstream Charts, which cannot be done by using values.yaml alone. For this reason, include the charts in our own repository. Normally, this means trouble keeping in sync with the upstream version. However, it is deprecated and archived, so should be okay for us at this moment. Longer term, we'll want to use their upcoming Kubernetes Operator.