fix: hardcode ES Helm volumeClaimTemplate label
Helm sets a helm.sh/chart label on EventStore Statefulset volumeClaimTemplates. The Helm app version was included in this label. This means that we could not upgrade the app version, as the StatefulSet had forbidden changes. For this reason, hardcode the helm.sh/chart label. Ideally, we would just remove the label, or at the least not include the version number. However, since we already have deployed a ES cluster in production with this exact label, we are forced to pin evenstore-0.2.5, even though this version will increase in the future