Upgrade to eventstore 20.10.2 (#32)
* Add new env variables to values.yaml, plus add init container

* update tgz

* remove securitycontext

* filepath

* build: check in ES Charts

EventStore is one of our dependencies for the registry-backend. We want to make changes to the upstream Charts, which cannot be done by using values.yaml alone. For this reason, include the charts in our own repository. Normally, this means trouble keeping in sync with the upstream version. However, it is deprecated and archived, so should be okay for us at this moment. Longer term, we'll want to use their upcoming Kubernetes Operator.

* feat: re-apply changes to ES charts

add initContainer and bugfix for running one node

* chore: bump ES Charts version

* build: update ES dependency for registry-backend

* fix: use int instead of float voor clusterSize comparison

Let users deploy with one node. The ES charts compared with 1.0, but that didn't work very nice

* fix: hardcode ES Helm volumeClaimTemplate label

Helm sets a helm.sh/chart label on EventStore Statefulset volumeClaimTemplates. The Helm app version was included in this label. This means that we could not upgrade the app version, as the StatefulSet had forbidden changes. For this reason, hardcode the helm.sh/chart label. Ideally, we would just remove the label, or at the least not include the version number. However, since we already have deployed a ES cluster in production with this exact label, we are forced to pin evenstore-0.2.5, even though this version will increase in the future

Co-authored-by: Stefan Bussemaker <Stefan.Bussemaker@kadaster.nl>